Post done, TODO: Post needs to be displayed in profile page, Custom adapter required

+++ app/src/main/java/models/HomeModel.kt
@@ -5,5 +5,5 @@ data class HomeModel (
     val img: String,
     val subTitle: String,
     val post: String,
-    val score: Double,
+    val score: Int,
 )
